% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartoon.R
\name{generate_cartoon_data}
\alias{generate_cartoon_data}
\title{Generate data for the retina cartoon}
\usage{
generate_cartoon_data(
  gene,
  dataset = "godzilla",
  label = T,
  color = "red",
  data_list = data_list
)
}
\arguments{
\item{gene}{Character string of the gene name for expression visualization. If the requested
gene is not found, it was most likely not annotated in the given dataset. Try adding the suffix
'-GRCh38' if you believe the gene may have been annotated in the more recent genome build.}

\item{dataset}{Character string of the dataset name. Options are: "godzilla", "retina_fov_vs_periph",
"retina_AIR_vs_control", "RPE_choroid_unselected", "RPE_choroid_CD31_selected", and
"CD31_choroid_infant_adult". Godzilla (a merged object of all retina+rpe+choroid
single-cell datasets published by 09/01/2020) is chosen by default. Names mirror datasets hosted on
Spectalce (https://oculargeneexpression.org/singlecell)}

\item{label}{TRUE/FALSE - should the cartoon have text labels for each cell type? True by default.}

\item{data_list}{the pre-generated list of cartoon outlines and expression matrices. This list
is pre-generated in data-raw and no adjustment to the parameter should be required. This
list has three elements (which are themselves lists): 1. a list of expression matrices. 2.
a list containing cartoon outlines WITHOUT labels. 3. a list containing cartoon outlines WITH
labels. The character name of the specified dataset is used to refer to the correct expression
matrix. The user-specified label parameter (see below) specified if the cartoon should include
labels or not.}
}
\description{
\code{generate_cartoon_data} returns a list with two elements. The first element is a
heatmap-cartoon genenerated by the grImport package. The second is a legend that related
expression to color. These elements can then be combined by the plot_cartoon function
to display a pretty cartoon plot
}
\examples{
\dontrun{
generate_cartoon_data(gene = "RHO", dataset = "godzilla")
generate_cartoon_data(gene = "RHO", dataset = "godzilla", label = T, color = "red", data_list = data_list)
}

}
