% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartoon.R
\name{plot_cartoon}
\alias{plot_cartoon}
\title{Plot the generated retina cartoon}
\usage{
plot_cartoon(cartoon_data)
}
\arguments{
\item{cartoon_data}{Cartoon data generated from the generate_cartoon_data. Cartoon data is
a list containing an XML object and an expression legend.}
}
\description{
\code{plot_cartoon} returns a list with two elements. The first element is a
heatmap-cartoon genenerated by the grImport package. The second is a legend that related
expression to color. These elements can then be combined by the plot_cartoon function
to display a pretty cartoon plot
}
\examples{
\dontrun{
generate_cartoon_data(gene = "RHO", dataset = "all_retina_rpe_chor") \%>\% plot_cartoon()
}

}
